kv (ticdc): remove unused code
