+++ cdc/kv/region_state.go
@@ -14,7 +14,6 @@
 package kv
 
 import (
-	"runtime"
 	"sync"
 
 	"github.com/pingcap/tiflow/cdc/kv/regionlock"
@@ -194,105 +193,4 @@ func (s *regionFeedState) getRegionInfo() regionInfo {
 
 func (s *regionFeedState) getRegionMeta() (uint64, tablepb.Span, string) {
 	return s.region.verID.GetID(), s.region.span, s.region.rpcCtx.Addr
-}
-
-type syncRegionFeedStateMap struct {
-	mu sync.RWMutex
-	// statesInternal is an internal field and must not be accessed from outside.
-	statesInternal map[uint64]*regionFeedState
-}
-
-func newSyncRegionFeedStateMap() *syncRegionFeedStateMap {
-	return &syncRegionFeedStateMap{
-		mu:             sync.RWMutex{},
-		statesInternal: make(map[uint64]*regionFeedState),
-	}
-}
-
-func (m *syncRegionFeedStateMap) iter(fn func(requestID uint64, state *regionFeedState) bool) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	for requestID, state := range m.statesInternal {
-		if !fn(requestID, state) {
-			break
-		}
-	}
-}
-
-func (m *syncRegionFeedStateMap) setByRequestID(requestID uint64, state *regionFeedState) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.statesInternal[requestID] = state
-}
-
-func (m *syncRegionFeedStateMap) setByRegionID(regionID uint64, state *regionFeedState) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.statesInternal[regionID] = state
-}
-
-func (m *syncRegionFeedStateMap) getByRegionID(regionID uint64) (*regionFeedState, bool) {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-	result, ok := m.statesInternal[regionID]
-	return result, ok
-}
-
-func (m *syncRegionFeedStateMap) delByRegionID(regionID uint64) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	delete(m.statesInternal, regionID)
-}
-
-type regionStateManagerInterface interface {
-	getState(regionID uint64) (*regionFeedState, bool)
-	setState(regionID uint64, state *regionFeedState)
-	delState(regionID uint64)
-}
-
-// regionStateManager provides the get/put way like a sync.Map, and it is divided
-// into several buckets to reduce lock contention
-type regionStateManager struct {
-	bucket int
-	states []*syncRegionFeedStateMap
-}
-
-func newRegionStateManager(bucket int) *regionStateManager {
-	if bucket <= 0 {
-		bucket = runtime.NumCPU()
-		if bucket > maxRegionStateBucket {
-			bucket = maxRegionStateBucket
-		}
-		if bucket < minRegionStateBucket {
-			bucket = minRegionStateBucket
-		}
-	}
-	rsm := &regionStateManager{
-		bucket: bucket,
-		states: make([]*syncRegionFeedStateMap, bucket),
-	}
-	for i := range rsm.states {
-		rsm.states[i] = newSyncRegionFeedStateMap()
-	}
-	return rsm
-}
-
-func (rsm *regionStateManager) getBucket(regionID uint64) int {
-	return int(regionID) % rsm.bucket
-}
-
-func (rsm *regionStateManager) getState(regionID uint64) (*regionFeedState, bool) {
-	bucket := rsm.getBucket(regionID)
-	state, ok := rsm.states[bucket].getByRegionID(regionID)
-	return state, ok
-}
-
-func (rsm *regionStateManager) setState(regionID uint64, state *regionFeedState) {
-	bucket := rsm.getBucket(regionID)
-	rsm.states[bucket].setByRegionID(regionID, state)
-}
-
-func (rsm *regionStateManager) delState(regionID uint64) {
-	bucket := rsm.getBucket(regionID)
-	rsm.states[bucket].delByRegionID(regionID)
 }
