+++ cdc/kv/shared_region_worker.go
@@ -127,20 +127,18 @@ func newResolvedTsBatch(ts uint64, stream *requestedStream) statefulEvent {
 }
 
 type sharedRegionWorker struct {
-	changefeed    model.ChangeFeedID
-	client        *SharedClient
-	statesManager *regionStateManager
-	inputCh       chan statefulEvent
-	metrics       *regionWorkerMetrics
+	changefeed model.ChangeFeedID
+	client     *SharedClient
+	inputCh    chan statefulEvent
+	metrics    *regionWorkerMetrics
 }
 
 func newSharedRegionWorker(c *SharedClient) *sharedRegionWorker {
 	return &sharedRegionWorker{
-		changefeed:    c.changefeed,
-		client:        c,
-		inputCh:       make(chan statefulEvent, regionWorkerInputChanSize),
-		statesManager: newRegionStateManager(-1),
-		metrics:       newRegionWorkerMetrics(c.changefeed, "shared", "shared"),
+		changefeed: c.changefeed,
+		client:     c,
+		inputCh:    make(chan statefulEvent, regionWorkerInputChanSize),
+		metrics:    newRegionWorkerMetrics(c.changefeed, "shared", "shared"),
 	}
 }
